Spark 버전별 Hive 소스 분기
